chore: remove "middleware" module, move all classes to mongo dir
